Adding .dockerignore file
This prevents files unnecessary to the Docker build from being included in the build context.  I'm not sure if it's best practice to exclude the Dockerfile or .dockerignore files themselves, I would have expected them to automatically be excluded, but Docker does seem to bring them in, so they're included here so that the build context is as small as can be (only the prepare-source.sh file is included in the build context).