Adding docker-compose.yml
This makes it slightly easier to run the container, as the required config values can be specified via environment variables on the host.  For example, the following command can be used:
APP_NAME=MyApp VERSION=0.0.1 SOURCE_DIR=/host/path/to/source OUTPUT_DIR=/desired/output/directory docker-compose up

Or if doing multiple builds, it might make sense to set the output directory environment variable first (or at login), rather than as part of the command:
export OUTPUT_DIR=/desired/output/directory
APP_NAME=MyApp VERSION=0.0.1 SOURCE_DIR=/host/path/to/source docker-compose up

This assumes you're happy for the output directories to be named SOURCES and SPECS, and to exist within the same parent directory, which is likely to be the case if you're about to treat that parent directory as the RPM build root.